Extract tree from analyze.
